+++ packages/tools/src/utilities/stackPrefetch/index.ts
@@ -1,12 +1,4 @@
-import {
-  enable,
-  disable,
-  setConfiguration,
-  getConfiguration,
-} from './stackPrefetch';
-
-import * as stackContextPrefetch from './stackContextPrefetch';
-
-const stackPrefetch = { enable, disable, setConfiguration, getConfiguration };
+import stackPrefetch from './stackPrefetch';
+import stackContextPrefetch from './stackContextPrefetch';
 
 export { stackPrefetch, stackContextPrefetch };
