+++ packages/tools/src/utilities/stackPrefetch/stackPrefetch.ts
@@ -9,10 +9,14 @@ import {
   getConfiguration as getCoreConfiguration,
 } from '@cornerstonejs/core';
 import { addToolState, getToolState } from './state';
-
-const requestType = Enums.RequestType.Prefetch;
-const priority = 0;
-const addToBeginning = true;
+import {
+  getStackData,
+  requestType,
+  priority,
+  getPromiseRemovedHandler,
+  nearestIndex,
+  range,
+} from './stackPrefetchUtils';
 
 let configuration = {
   maxImagesToPrefetch: Infinity,
@@ -21,66 +25,6 @@ let configuration = {
 
 let resetPrefetchTimeout;
 const resetPrefetchDelay = 10;
-
-function range(lowEnd, highEnd) {
-  // Javascript version of Python's range function
-  // http://stackoverflow.com/questions/3895478/does-javascript-have-a-method-like-range-to-generate-an-array-based-on-suppl
-  lowEnd = Math.round(lowEnd) || 0;
-  highEnd = Math.round(highEnd) || 0;
-
-  const arr = [];
-  let c = highEnd - lowEnd + 1;
-
-  if (c <= 0) {
-    return arr;
-  }
-
-  while (c--) {
-    arr[c] = highEnd--;
-  }
-
-  return arr;
-}
-
-function nearestIndex(arr, x) {
-  // Return index of nearest values in array
-  // http://stackoverflow.com/questions/25854212/return-index-of-nearest-values-in-an-array
-  let low = 0;
-  let high = arr.length - 1;
-
-  arr.forEach((v, idx) => {
-    if (v < x) {
-      low = Math.max(idx, low);
-    } else if (v > x) {
-      high = Math.min(idx, high);
-    }
-  });
-
-  return { low, high };
-}
-
-function getStackData(element) {
-  const enabledElement = getEnabledElement(element);
-
-  if (!enabledElement) {
-    throw new Error(
-      'stackPrefetch: element must be a valid Cornerstone enabled element'
-    );
-  }
-
-  const { viewport } = enabledElement;
-
-  if (!(viewport instanceof StackViewport)) {
-    throw new Error(
-      'stackPrefetch: element must be a StackViewport, VolumeViewport stackPrefetch not yet implemented'
-    );
-  }
-
-  return {
-    currentImageIdIndex: viewport.getCurrentImageIdIndex(),
-    imageIds: viewport.getImageIds(),
-  };
-}
 
 function prefetch(element) {
   // Get the stackPrefetch tool data
@@ -97,6 +41,8 @@ function prefetch(element) {
     console.warn('CornerstoneTools.stackPrefetch: No images in stack.');
     return;
   }
+
+  const { currentImageIdIndex } = stack;
 
   // If all the requests are complete, disable the stackPrefetch tool
   if (
@@ -134,9 +80,20 @@ function prefetch(element) {
       return;
     }
 
-    const imageLoadObject = cache.getImageLoadObject(imageId);
+    const distance = Math.abs(currentImageIdIndex - imageIdIndex);
+    // For nearby objects, ensure the last accessed time is updated
+    // by using getImageLoadObject.
+    // For more distant objects, just check if available, but dont
+    // change the access time.
+    // This allows throwing data that hasn't been accessed and is not
+    // nearby.
+    const imageCached =
+      distance < 6
+        ? cache.getImageLoadObject(imageId)
+        : cache.isImageIdCached(imageId);
 
-    if (imageLoadObject) {
+    if (imageCached) {
+      // Already in cache
       removeFromList(imageIdIndex);
     }
   });
@@ -167,40 +124,7 @@ function prefetch(element) {
     const imageIdIndex = stack.imageIds.indexOf(image.imageId);
 
     removeFromList(imageIdIndex);
-
-    // triggerEvent(element, EVENTS.STACK_PREFETCH_IMAGE_LOADED, {
-    //   element,
-    //   imageId: image.imageId,
-    //   imageIndex: imageIdIndex,
-    //   stackPrefetch,
-    //   stack,
-    // });
-
-    // If there are no more images to fetch
-    // if (
-    //   !(
-    //     stackPrefetch.indicesToRequest &&
-    //     stackPrefetch.indicesToRequest.length > 0
-    //   )
-    // ) {
-    //   triggerEvent(element, EVENTS.STACK_PREFETCH_DONE, {
-    //     element,
-    //     stackPrefetch,
-    //     stack,
-    //   });
-    // }
   }
-
-  // Retrieve the errorLoadingHandler if one exists
-  // const errorLoadingHandler =
-  //   loadHandlerManager.getErrorLoadingHandler(element);
-
-  // function failCallback(error) {
-  //   logger.log('prefetch errored: %o', error);
-  //   if (errorLoadingHandler) {
-  //     errorLoadingHandler(element, imageId, error, 'stackPrefetch');
-  //   }
-  // }
 
   // Prefetch images around the current image (before and after)
   let lowerIndex = nearest.low;
@@ -269,49 +193,6 @@ function prefetch(element) {
       // addToBeginning
     );
   });
-}
-
-function getPromiseRemovedHandler(element) {
-  return function (e) {
-    const eventData = e.detail;
-
-    // When an imagePromise has been pushed out of the cache, re-add its index
-    // It to the indicesToRequest list so that it will be retrieved later if the
-    // CurrentImageIdIndex is changed to an image nearby
-    let stackData;
-
-    try {
-      // It will throw an exception in some cases (eg: thumbnails)
-      stackData = getStackData(element);
-    } catch (error) {
-      return;
-    }
-
-    if (!stackData || !stackData.imageIds || stackData.imageIds.length === 0) {
-      return;
-    }
-
-    const stack = stackData;
-    const imageIdIndex = stack.imageIds.indexOf(eventData.imageId);
-
-    // Make sure the image that was removed is actually in this stack
-    // Before adding it to the indicesToRequest array
-    if (imageIdIndex < 0) {
-      return;
-    }
-
-    const stackPrefetchData = getToolState(element);
-
-    if (
-      !stackPrefetchData ||
-      !stackPrefetchData.data ||
-      !stackPrefetchData.data.length
-    ) {
-      return;
-    }
-
-    stackPrefetchData.indicesToRequest.push(imageIdIndex);
-  };
 }
 
 function onImageUpdated(e) {
@@ -402,4 +283,6 @@ function setConfiguration(config) {
   configuration = config;
 }
 
-export { enable, disable, getConfiguration, setConfiguration };
+const stackPrefetch = { enable, disable, getConfiguration, setConfiguration };
+
+export default stackPrefetch;
